refactor(evm): move assembler peepholes
