+++ crates/codegen/src/backend/evm/mod.rs
@@ -12,8 +12,7 @@ pub use codegen::{EvmArtifact, EvmCodegen, EvmCodegenConfig};
 pub mod assembler;
 pub use assembler::{AssembledCode, Assembler, AssemblerConfig, Label};
 
-pub mod peephole;
-pub use peephole::PeepholeOptimizer;
+mod peephole;
 
 pub mod stack;
 pub use stack::{SpillManager, SpillSlot, StackModel, StackScheduler};
